Fix for crashes in portrait mode + Conversion of Linear layout to Constraint layout

+++ app/src/main/java/com/quickhandslogistics/utils/ValueUtils.kt
@@ -41,14 +41,14 @@ object ValueUtils {
 
     fun getHoursFromMinutes(minutes:String?) : String {
         minutes?.let {
-            return (it.toDouble() / 60).toString()
+            return (it.toDouble() / 60).toInt().toString()
         }
         return "0"
     }
 
     fun getRemainingMinutes(minutes:String?) : String {
         minutes?.let {
-            return (it.toDouble() % 60).toString()
+            return (it.toDouble() % 60).toInt().toString()
         }
         return "0"
     }
